Application now provides a GuiPreferences & from preferences() PreferencesDialogue now modifies those in-place. Also, the Apply and Save button both work as expected now. Other places that use Application::preferences() updated accordingly.
Application is Qt-specific, and only for the GUI application, therefore
the preferences it stores are always GuiPreferences. Therefore, it now
stores a GuiPreferences object, and does so on the stack. It is now
guaranteed to provide a valid object from preferences().

See the list of modified files for where the change to
Application::preferences() has had its impact.