Implemented fully-working completion for source and destination.
Completion uses setRootPath() on the QFileSystemModel to ensure that
completions are possible when editing a partial path. Otherwise,
completions only take effect when entering a new path from scratch.
Having read the Qt docs for this, it seems like this is not what the
setRootPath() method is for but I've yet to find anything else that
works. Perhaps the docs are just not clear enough for me...